backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
python-decouple==3.8
psycopg2-binary==2.9.11
gunicorn==21.2.0
whitenoise==6.6.0
PuLP>=2.7.0
pytest>=7.0.0
pytest-django>=4.5.0
Pillow==10.4.0
django-channels==4.1.0
channels-redis==4.2.0
django-redis==5.4.0
redis==5.0.1

# Socket.IO dependencies
python-socketio>=5.8.0
eventlet>=0.33.0